Using more appropriate types

+++ src/main/java/com/frogrilla/echoes/common/block/CorruptedEchoRodBlock.java
@@ -34,7 +34,7 @@ public class CorruptedEchoRodBlock extends EchoRodBlock{
             if(powered){
                 doEffects((ServerWorld) world, pos, state.get(FACING));
                 SignalManager manager = PersistentManagerState.getServerWorldState((ServerWorld) world).signalManager;
-                CorruptedSignal signal = new CorruptedSignal(pos.offset(state.get(FACING), 2), state.get(FACING), power);
+                CorruptedSignal signal = new CorruptedSignal(pos.offset(state.get(FACING), 2), state.get(FACING), (byte) power);
                 manager.addSignal(signal);
             }
             world.setBlockState(pos, state.with(POWERED, powered));
